+++ oasis-web/src/typings/elegant-router.d.ts
@@ -33,7 +33,7 @@ declare module "@elegant-router/types" {
     "schedule": "/schedule";
     "schedule_dispatch": "/schedule/dispatch";
     "schedule_job": "/schedule/job";
-    "schedule_job_alarm-events": "/schedule/job/alarm-events/:jobId";
+    "schedule_job_alarm-events": "/schedule/job/alarm-events";
     "schedule_log": "/schedule/log";
   };